Serialization format does not need lengths
Since the tag generated is a fixed 16 bits, the ciphertext payload can
just be put at the end of the token and have its size be implicit.

This changes one of the tests' returns from a bad format error to an
invalid mac error, since the error is caught at a later point.

Also changed two type signatures that accepted keys where encoded keys
were expected.